Initial commit for the cookbook
The DataSourceElement is a new module for mapbender to access and/or manipulate different kinds of datasources.
Currently there is only an api documentation and it can be challenging for new developers to get started with it.
To remove this problem the new cookbook should give a short overview about this element and allow developer a quick and easy start
with the DataSourceElement.